test config refactoring
